+++ plotoptix/utils.py
@@ -10,6 +10,67 @@ from plotoptix._load_lib import load_optix
 
 _optix = load_optix()
 
+
+def _make_contiguous_vector(a: Optional[Any], n_dim: int) -> Optional[np.ndarray]:
+    if a is None: return None
+
+    if not isinstance(a, np.ndarray) or (a.dtype != np.float32):
+        a = np.ascontiguousarray(a, dtype=np.float32)
+    if len(a.shape) > 1: a = a.flatten()
+    if a.shape[0] > n_dim: a = a[:n_dim]
+    if a.shape[0] == 1: a = np.full(n_dim, a[0], dtype=np.float32)
+    if a.shape[0] < n_dim: return None
+
+    if not a.flags['C_CONTIGUOUS']: a = np.ascontiguousarray(a, dtype=np.float32)
+
+    return a
+
+def _make_contiguous_3d(a: Optional[Any], n: int = -1, extend_scalars = False) -> Optional[np.ndarray]:
+    if a is None: return None
+
+    if not isinstance(a, np.ndarray): a = np.ascontiguousarray(a, dtype=np.float32)
+
+    if a.dtype != np.float32: a = np.ascontiguousarray(a, dtype=np.float32)
+
+    if len(a.shape) == 1:
+        if a.shape[0] == 1: a = np.full((1, 3), a[0], dtype=np.float32)
+        elif a.shape[0] == n: a = np.reshape(a, (n, 1))
+        elif a.shape[0] == 3: a = np.reshape(a, (1, 3))
+        else:
+            raise ValueError("Input shape not matching single 3D vector nor desired array length.")
+
+    if len(a.shape) > 2:
+        m = functools.reduce(operator.mul, a.shape[:-1], 1)
+        if (n >= 0) and (n != m):
+            raise ValueError("Input shape not matching desired array length.")
+        a = np.reshape(a, (m, a.shape[-1]))
+
+    if n >= 0:
+        if (a.shape[0] == 1) and (n != a.shape[0]):
+            _a = np.zeros((n, a.shape[-1]), dtype=np.float32)
+            _a[:] = a[0]
+            a = _a
+        if n != a.shape[0]:
+            raise ValueError("Input shape not matching desired array length.")
+
+    if a.shape[-1] != 3:
+        _a = np.zeros((a.shape[0], 3), dtype=np.float32)
+        if a.shape[-1] == 1:
+            if extend_scalars:
+                _a[:,0] = a[:,0]
+                _a[:,1] = a[:,0]
+                _a[:,2] = a[:,0]
+            else:
+                _a[:,0] = a[:,0]
+        elif a.shape[-1] == 2: _a[:,[0,1]] = a[:,[0,1]]
+        else: _a[:,[0,1,2]] = a[:,[0,1,2]]
+        a = _a
+
+    if not a.flags['C_CONTIGUOUS']: a = np.ascontiguousarray(a, dtype=np.float32)
+
+    return a
+
+
 def map_to_colors(x: Any, cm_name: str) -> np.ndarray:
     """Map input variable to matplotlib color palette.